fix: preserve mating_day int16 dtype and fix eat_food_vectorized food conservation
- np.where(is_female, mating_days, -99) silently promoted int16 to int64;
  now uses np.int16(-99) and .astype(np.int16) to preserve dtype
- NumPy fallback in eat_food_vectorized had food conservation bug:
  actual_available was computed from already-depleted grid. Rewrote to
  compute total demand per cell, cap at available food, then distribute
  proportional shares to agents
- Added 3 tests in TestDtypeAndFoodBugs

+++ src/cenop/agents/population.py
@@ -840,7 +840,7 @@ class PorpoisePopulation:
         # Mating day (females only, N(225, 20))
         mating_days = np.clip(self.rng.normal(225, 20, self.count), 0, 359).astype(np.int16)
         # Apply only to females, others stay -99
-        self.mating_day = np.where(self.is_female, mating_days, -99)
+        self.mating_day = np.where(self.is_female, mating_days, np.int16(-99)).astype(np.int16)
 
         # Initialize pregnancy state (Java Porpoise.java:165-178)
         maturity_age = self.params.maturity_age
